Revert SessionListener rename for BC
SessionListener and TestSessionListener are extended outside of Symfony
in Silex. These classes cannot be renamed.
