+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/test.xml
@@ -20,7 +20,7 @@
 
         <service id="test.client.cookiejar" class="Symfony\Component\BrowserKit\CookieJar" shared="false" />
 
-        <service id="test.session.listener" class="Symfony\Component\HttpKernel\EventListener\TestSessionListener">
+        <service id="test.session.listener" class="Symfony\Component\HttpKernel\EventListener\ContainerAwareTestSessionListener">
             <tag name="kernel.event_subscriber" />
             <tag name="container.service_subscriber" id="session" />
             <argument type="service" id="container" />
