+++ src/Symfony/Component/HttpKernel/EventListener/ContainerAwareSessionListener.php
@@ -1,0 +1,52 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\HttpKernel\EventListener;
+
+use Psr\Container\ContainerInterface;
+use Symfony\Component\DependencyInjection\ServiceSubscriberInterface;
+use Symfony\Component\HttpFoundation\Session\SessionInterface;
+
+/**
+ * Sets the session in the request.
+ *
+ * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @final since version 3.3
+ */
+class ContainerAwareSessionListener extends SessionListener implements ServiceSubscriberInterface
+{
+    private $container;
+
+    public function __construct(ContainerInterface $container)
+    {
+        $this->container = $container;
+    }
+
+    protected function getSession()
+    {
+        if (!$this->container->has('session')) {
+            return;
+        }
+
+        return $this->container->get('session');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public static function getSubscribedServices()
+    {
+        return array(
+            'session' => '?'.SessionInterface::class,
+        );
+    }
+}
